Backend/requirements.txt
Django==4.1
django-allauth==0.51.0
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-rest-framework==0.1.0
django-taggit==3.1.0
django-taggit-serializer==0.1.7
djangorestframework==3.14.0
Pillow==9.2.0
psycopg2==2.9.5
PyJWT==2.5.0
